fix(api): clear layer-import-guard violations
The layer guard rejects api -> services and services -> api edges. After
the 10D/10G migration two classes of violation surfaced:

* Every router carried service: XService = Depends(get_x_service)
  annotations that pulled the orchestrator classes into the api layer
  for type info. Drop the annotations and the corresponding
  from services.orchestrators.x import X imports — runtime behaviour
  is unchanged, FastAPI does not need the class hint to resolve the
  dependency.

* UserCreate / UserUpdate / UserPublic are shared between
  the chat-completion router (request body) and UserService
  (constructor input). They had to live somewhere both layers can see
  without crossing the api↔services boundary, so move the canonical
  definitions into core/models/user.py next to the existing
  User / UserPartition / ApiKey domain models, and have
  api/schemas/admin/users.py re-export them for HTTP-layer callers.
