+++ openrag/api/routers/admin/workspaces.py
@@ -16,7 +16,6 @@ from api.dependencies.auth import require_partition_editor, require_partition_ow
 from di.providers import get_workspace_service
 from fastapi import APIRouter, Depends, HTTPException, status
 from pydantic import BaseModel, ConfigDict, field_validator
-from services.orchestrators.workspace_service import WorkspaceService
 from utils.logger import get_logger
 
 router = APIRouter()
@@ -50,7 +49,7 @@ class AddFilesRequest(BaseModel):
 async def require_workspace_in_partition(
     partition: str,
     workspace_id: str,
-    service: WorkspaceService = Depends(get_workspace_service),
+    service=Depends(get_workspace_service),
 ) -> dict:
     """Validate that a workspace exists and belongs to the given partition."""
     ws = await service.get_workspace(workspace_id)
@@ -67,7 +66,7 @@ async def create_workspace(
     partition: str,
     body: CreateWorkspaceRequest,
     user=Depends(require_partition_editor),
-    service: WorkspaceService = Depends(get_workspace_service),
+    service=Depends(get_workspace_service),
 ):
     if await service.get_workspace(body.workspace_id):
         raise HTTPException(
@@ -89,7 +88,7 @@ async def create_workspace(
 )
 async def list_workspaces(
     partition: str,
-    service: WorkspaceService = Depends(get_workspace_service),
+    service=Depends(get_workspace_service),
 ):
     return {"workspaces": await service.list_workspaces(partition)}
 
@@ -110,7 +109,7 @@ async def delete_workspace(
     partition: str,
     workspace_id: str,
     _ws=Depends(require_workspace_in_partition),
-    service: WorkspaceService = Depends(get_workspace_service),
+    service=Depends(get_workspace_service),
 ):
     result = await service.delete_workspace(partition, workspace_id)
     return {"status": "deleted", **result}
@@ -125,7 +124,7 @@ async def add_files_to_workspace(
     workspace_id: str,
     body: AddFilesRequest,
     _ws=Depends(require_workspace_in_partition),
-    service: WorkspaceService = Depends(get_workspace_service),
+    service=Depends(get_workspace_service),
 ):
     existing_ids = await service.get_existing_file_ids(partition, body.file_ids)
     unknown_ids = sorted(set(body.file_ids) - set(existing_ids))
@@ -151,7 +150,7 @@ async def add_files_to_workspace(
 async def list_workspace_files(
     workspace_id: str,
     _ws=Depends(require_workspace_in_partition),
-    service: WorkspaceService = Depends(get_workspace_service),
+    service=Depends(get_workspace_service),
 ):
     return {"file_ids": await service.list_files(workspace_id)}
 
@@ -163,7 +162,7 @@ async def list_workspace_files(
 async def list_file_workspaces(
     partition: str,
     file_id: str,
-    service: WorkspaceService = Depends(get_workspace_service),
+    service=Depends(get_workspace_service),
 ):
     workspace_ids = await service.get_file_workspaces(file_id, partition)
     return {"file_id": file_id, "workspace_ids": workspace_ids}
@@ -177,7 +176,7 @@ async def remove_file_from_workspace(
     workspace_id: str,
     file_id: str,
     _ws=Depends(require_workspace_in_partition),
-    service: WorkspaceService = Depends(get_workspace_service),
+    service=Depends(get_workspace_service),
 ):
     removed = await service.remove_file(workspace_id, file_id)
     if not removed:
