+++ openrag/api/schemas/admin/users.py
@@ -1,24 +1,10 @@
-from pydantic import BaseModel, ConfigDict, Field
+"""User request/response schemas.
 
+Re-exports the request DTOs from :mod:`core.models.user` so the
+:class:`UserService` orchestrator and the FastAPI router both reference
+the same Pydantic models without the service layer reaching into ``api/``.
+"""
 
-class UserBase(BaseModel):
-    display_name: str | None = None
-    external_user_id: str | None = None
-    email: str | None = Field(default=None, examples=[None])
-    is_admin: bool = False
-    file_quota: int | None = Field(default=10)
+from core.models.user import UserBase, UserCreate, UserPublic, UserUpdate
 
-
-class UserCreate(UserBase):
-    model_config = ConfigDict(extra="allow")
-
-
-class UserUpdate(UserBase):
-    model_config = ConfigDict(extra="allow")
-
-
-class UserPublic(UserBase):
-    id: int
-    created_at: str | None
-    file_quota: int | None = None
-    file_count: int | None = None
+__all__ = ["UserBase", "UserCreate", "UserPublic", "UserUpdate"]
