+++ openrag/services/orchestrators/test_user_service.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 
 import pytest
-from api.schemas.admin.users import UserCreate, UserUpdate
-from core.models.user import User
+from core.models.user import User, UserCreate, UserUpdate
 from core.utils.exceptions import UserNotFoundError, ValidationError
 from services.orchestrators.user_service import UserService
